+++ C++/data/hash/hmac.h
@@ -24,8 +24,8 @@ namespace hash
 	public:
 		HMAC(const uint8_t* key, std::size_t ksize)
 		{
-			std::fill(ki, ki + H::block_size, 0x36);
-			std::fill(ko, ko + H::block_size, 0x5c);
+			std::fill(ki, ki + H::block_size, uint8_t(0x36));
+			std::fill(ko, ko + H::block_size, uint8_t(0x5c));
 			if(ksize <= H::block_size)
 			{
 				for(uint_fast8_t i = 0; i < ksize; ++i)
